+++ study/api-src/org/labkey/api/specimen/SpecimenTypeBeanProperty.java
@@ -1,0 +1,32 @@
+package org.labkey.api.specimen;
+
+import org.labkey.api.query.FieldKey;
+
+public class SpecimenTypeBeanProperty
+{
+    private final FieldKey _typeKey;
+    private final String _beanProperty;
+    private final SpecimenTypeLevel _level;
+
+    public SpecimenTypeBeanProperty(FieldKey typeKey, String beanProperty, SpecimenTypeLevel level)
+    {
+        _typeKey = typeKey;
+        _beanProperty = beanProperty;
+        _level = level;
+    }
+
+    public FieldKey getTypeKey()
+    {
+        return _typeKey;
+    }
+
+    public String getBeanProperty()
+    {
+        return _beanProperty;
+    }
+
+    public SpecimenTypeLevel getLevel()
+    {
+        return _level;
+    }
+}
