+++ study/api-src/org/labkey/api/specimen/SpecimenWebPart.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2014-2019 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.api.specimen;
+
+import org.labkey.api.specimen.settings.SettingsManager;
+import org.labkey.api.study.Study;
+import org.labkey.api.view.JspView;
+
+/**
+ * User: Mark Igra
+ * Date: Jul 28, 2006
+ * Time: 10:50:05 AM
+ */
+public class SpecimenWebPart extends JspView<SpecimenWebPart.SpecimenWebPartBean>
+{
+    @SuppressWarnings("unused") // Invoked via reflection
+    public SpecimenWebPart()
+    {
+        this(false, null);
+    }
+
+    public SpecimenWebPart(boolean wide, Study study)
+    {
+        super("/org/labkey/study/view/specimen/webPart.jsp", new SpecimenWebPartBean(wide,
+                (null != study) && SettingsManager.get().getRepositorySettings(study.getContainer()).isSpecimenDataEditable()));
+        setTitle("Specimens");
+    }
+
+    public static class SpecimenWebPartBean
+    {
+        private final boolean _wide;
+        private final boolean _isEditableSpecimens;
+
+        public SpecimenWebPartBean(boolean wide, boolean isEditableSpecimens)
+        {
+            _wide = wide;
+            _isEditableSpecimens = isEditableSpecimens;
+        }
+
+        public boolean isWide()
+        {
+            return _wide;
+        }
+
+        public boolean isEditableSpecimens()
+        {
+            return _isEditableSpecimens;
+        }
+    }
+}
