+++ study/src/org/labkey/study/specimen/notifications/DefaultRequestNotification.java
@@ -30,8 +30,8 @@ import org.labkey.api.specimen.notifications.NotificationRecipientSet;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
 import org.labkey.api.specimen.settings.RequestNotificationSettings;
+import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.view.ViewContext;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.query.SpecimenQueryView;
 
 import java.io.IOException;
@@ -81,7 +81,7 @@ public class DefaultRequestNotification
 
     private void addSpecimenListFileIfNeeded(ViewContext context) throws Exception
     {
-        RequestNotificationSettings settings = SpecimenManager.getInstance().getRequestNotificationSettings(_request.getContainer());
+        RequestNotificationSettings settings = SettingsManager.get().getRequestNotificationSettings(_request.getContainer());
         if (RequestNotificationSettings.SpecimensAttachmentEnum.ExcelAttachment == settings.getSpecimensAttachmentEnum() ||
             RequestNotificationSettings.SpecimensAttachmentEnum.TextAttachment == settings.getSpecimensAttachmentEnum())
         {
