+++ study/src/org/labkey/study/specimen/report/SpecimenTypeVisitReport.java
@@ -16,13 +16,13 @@
 package org.labkey.study.specimen.report;
 
 import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.specimen.SpecimenTypeLevel;
+import org.labkey.api.specimen.report.SummaryByVisitType;
 import org.labkey.api.util.DemoMode;
 import org.labkey.api.util.Formats;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.study.SpecimenManager;
-import org.labkey.study.SpecimenManager.SpecimenTypeLevel;
-import org.labkey.api.specimen.report.SummaryByVisitType;
 import org.labkey.study.controllers.specimen.SpecimenController;
 import org.labkey.study.model.VisitImpl;
