+++ study/src/org/labkey/study/specimen/report/SpecimenVisitReport.java
@@ -22,6 +22,7 @@ import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryDefinition;
 import org.labkey.api.query.QueryService;
 import org.labkey.api.security.User;
+import org.labkey.api.specimen.SpecimenTypeLevel;
 import org.labkey.api.specimen.report.SpecimenReportCellData;
 import org.labkey.api.specimen.report.SpecimenReportTitle;
 import org.labkey.api.study.Study;
@@ -30,7 +31,6 @@ import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.study.CohortFilter;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.model.ParticipantGroup;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.StudyManager;
@@ -395,7 +395,7 @@ public abstract class SpecimenVisitReport<CELLDATA extends SpecimenReportCellDat
         return _parameters.getBaseCustomViewName();
     }
 
-    protected SpecimenManager.SpecimenTypeLevel getTypeLevelEnum()
+    protected SpecimenTypeLevel getTypeLevelEnum()
     {
         return _parameters.getTypeLevelEnum();
     }
